Data import/export: added number prefix to exported files; moved modules into dataExport folder
