+++ server/job/jobCreator.js
@@ -9,7 +9,7 @@ import CollectImportJob from '@server/modules/collectImport/service/collectImpor
 import CollectDataImportJob from '@server/modules/collectImport/service/collectImport/collectDataImportJob'
 import DataImportJob from '@server/modules/dataImport/service/DataImportJob'
 import DataImportValidationJob from '@server/modules/dataImport/service/DataImportValidationJob'
-import ExportCsvDataJob from '@server/modules/survey/service/export/exportCsvDataJob'
+import DataExportJob from '@server/modules/dataExport/service/DataExportJob'
 import RecordsCloneJob from '@server/modules/record/service/recordsCloneJob'
 import SurveyCloneJob from '@server/modules/survey/service/clone/surveyCloneJob'
 import SurveyExportJob from '@server/modules/survey/service/surveyExport/surveyExportJob'
@@ -29,7 +29,7 @@ const jobClasses = [
   CollectDataImportJob,
   DataImportJob,
   DataImportValidationJob,
-  ExportCsvDataJob,
+  DataExportJob,
   RecordsCloneJob,
   SurveyCloneJob,
   SurveyExportJob,
