+++ server/modules/dataExport/api/dataExportApi.js
@@ -8,10 +8,9 @@ import * as JobUtils from '@server/job/jobUtils'
 
 import * as AuthMiddleware from '@server/modules/auth/authApiMiddleware'
 import * as SurveyService from '@server/modules/survey/service/surveyService'
+import * as DataExportService from '../service/dataExportService'
 
 export const init = (app) => {
-  // export-csv-data
-  // generate zip with CSV
   app.post(
     '/survey/:surveyId/data-export/csv',
     AuthMiddleware.requireRecordsExportPermission,
@@ -33,7 +32,7 @@ export const init = (app) => {
 
         const user = Request.getUser(req)
 
-        const job = SurveyService.startExportCsvDataJob({
+        const job = DataExportService.startCsvDataExportJob({
           user,
           surveyId,
           cycle,
