+++ server/modules/dataExport/service/DataExportJob/index.js
@@ -1,0 +1,1 @@
+export { default } from './DataExportJob'
