+++ server/modules/record/api/recordApi.js
@@ -20,7 +20,6 @@ import {
   requireRecordViewPermission,
   requireRecordsEditPermission,
 } from '../../auth/authApiMiddleware'
-import { DataImportTemplateService } from '@server/modules/dataImport/service/dataImportTemplateService'
 
 export const init = (app) => {
   // ==== CREATE
@@ -67,61 +66,6 @@ export const init = (app) => {
       await RecordService.persistNode({ socketId, user, surveyId, cycle, draft, node, file, timezoneOffset })
 
       sendOk(res)
-    } catch (error) {
-      next(error)
-    }
-  })
-
-  app.post('/survey/:surveyId/record/importfromcollect', requireRecordCreatePermission, async (req, res, next) => {
-    try {
-      const user = Request.getUser(req)
-      const { surveyId, deleteAllRecords, cycle, forceImport } = Request.getParams(req)
-      const file = Request.getFile(req)
-
-      const job = RecordService.startCollectDataImportJob({
-        user,
-        surveyId,
-        filePath: file.tempFilePath,
-        deleteAllRecords,
-        cycle,
-        forceImport,
-      })
-      const jobSerialized = JobUtils.jobToJSON(job)
-      res.json({ job: jobSerialized })
-    } catch (error) {
-      next(error)
-    }
-  })
-
-  app.post('/survey/:surveyId/record/importfromcsv', requireRecordCreatePermission, async (req, res, next) => {
-    try {
-      const user = Request.getUser(req)
-      const {
-        surveyId,
-        cycle,
-        nodeDefUuid,
-        dryRun,
-        insertNewRecords,
-        insertMissingNodes,
-        updateRecordsInAnalysis,
-        abortOnErrors,
-      } = Request.getParams(req)
-      const filePath = Request.getFilePath(req)
-
-      const job = RecordService.startCSVDataImportJob({
-        user,
-        surveyId,
-        filePath,
-        cycle,
-        nodeDefUuid,
-        dryRun,
-        insertNewRecords,
-        insertMissingNodes,
-        updateRecordsInAnalysis,
-        abortOnErrors,
-      })
-      const jobSerialized = JobUtils.jobToJSON(job)
-      res.json({ job: jobSerialized })
     } catch (error) {
       next(error)
     }
@@ -300,26 +244,6 @@ export const init = (app) => {
       const { surveyId, cycle, lang, recordUuid } = Request.getParams(req)
 
       await RecordService.exportValidationReportToCSV({ res, surveyId, cycle, lang, recordUuid })
-    } catch (error) {
-      next(error)
-    }
-  })
-
-  app.get('/survey/:surveyId/record/importfromcsv/template', requireRecordCreatePermission, async (req, res, next) => {
-    try {
-      const { surveyId, nodeDefUuid, cycle } = Request.getParams(req)
-
-      await DataImportTemplateService.exportDataImportTemplate({ surveyId, cycle, nodeDefUuid, res })
-    } catch (error) {
-      next(error)
-    }
-  })
-
-  app.get('/survey/:surveyId/record/importfromcsv/templates', requireRecordCreatePermission, async (req, res, next) => {
-    try {
-      const { surveyId, cycle } = Request.getParams(req)
-
-      await DataImportTemplateService.exportAllDataImportTemplates({ surveyId, cycle, res })
     } catch (error) {
       next(error)
     }
