+++ server/system/apiRouter.js
@@ -11,6 +11,7 @@ import * as collectImportApi from '@server/modules/collectImport/api/collectImpo
 import * as arenaImportApi from '@server/modules/arenaImport/api/arenaImportApi'
 import * as recordApi from '@server/modules/record/api/recordApi'
 import * as dataExportApi from '@server/modules/dataExport/api/dataExportApi'
+import * as dataImportApi from '@server/modules/dataImport/api/dataImportApi'
 import * as surveyRdbApi from '@server/modules/surveyRdb/api/surveyRdbApi'
 import * as reportingApi from '@server/modules/reporting/api'
 import * as analysisApi from '@server/modules/analysis/api'
@@ -37,6 +38,7 @@ arenaImportApi.init(router)
 
 recordApi.init(router)
 dataExportApi.init(router)
+dataImportApi.init(router)
 
 surveyRdbApi.init(router)
 reportingApi.init(router)
